+++ src/Ext/TechnoType/Body.h
@@ -260,6 +260,13 @@ public:
 		Valueable<bool> KeepTargetOnMove;
 		Valueable<Leptons> KeepTargetOnMove_ExtraDistance;
 
+		Nullable<bool> DirectionalArmor {};
+		Nullable<float> DirectionalArmor_FrontMultiplier {};
+		Nullable<float> DirectionalArmor_SideMultiplier {};
+		Nullable<float> DirectionalArmor_BackMultiplier {};
+		Nullable<float> DirectionalArmor_FrontField {};
+		Nullable<float> DirectionalArmor_BackField {};
+
 		Valueable<int> Power;
 
     	Nullable<UnitTypeClass*> Image_ConditionYellow;